Making separate roll dialog class
