+++ module/documents/combat.mjs
@@ -1,10 +1,11 @@
 import { Dice } from "../dice.mjs";
 import { E20 } from "../helpers/config.mjs";
+import { RollDialog } from "../helpers/roll-dialog.mjs";
 
 export class Essence20Combat extends Combat {
   constructor(data, context) {
     super(data, context);
-    this._dice = new Dice(E20, ChatMessage);
+    this._dice = new Dice(E20, ChatMessage, new RollDialog(CONFIG.E20));
   }
 
   /**
